agent/requirements.txt
livekit
livekit-agents
livekit-plugins-google>=0.6.0
python-dotenv

